+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### Version 2.0
+* update support for WP version 6.0.2
+* add support for episode type, featured image, podcasts, tags, iTunes season number, iTunes episode number, and iTunes episode type 
+* update instructions for formatting data.
+* leading and trailing whitespace will be trimmed from submitted data
+* various code improvements
+* implement bug fix and code improvements by @tkittich https://profiles.wordpress.org/tkittich/
+
 ### Version 1.0
 
 * Initial release